Improve cv::Rect documentation
The documentation seems incomplete as x & y origins are to be calculated from bottom-left in algorithms than top-left.
Adding this line helps in clearifying the meaning of x & y in terms of algorithms.
Documentation in code provides quick access to definitions rather than going online and searching through the official documentation.